EHN: Make the console program importable by other program
